requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.6.1
python-dotenv>=1.0.0
pydantic>=2.5.2
rich>=13.7.0
typer>=0.9.0
sqlalchemy>=2.0.25
